+++ CHANGELOG.md
@@ -25,6 +25,8 @@ This changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.
 - `alias`, `staked_account_id`, `staked_node_id` and `decline_staking_reward` fields to AccountCreateTransaction
 - `staked_account_id`, `staked_node_id` and `decline_staking_reward` fields to AccountInfo
 - Added `examples/token_create_transaction_supply_key.py` to demonstrate token creation with and without a supply key.
+- Added BatchTransaction class
+
 
 ### Changed